Fix a TypeError
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/home/user/python-mingus-master/unittest/test_fft.py", line 22, in test_find_melody
    fft.find_melody('440_880_clean.wav', 512)[:2])
  File "../mingus/extra/fft.py", line 177, in find_melody
    for d in analyze_chunks(data, freq, bits, chunksize):
  File "../mingus/extra/fft.py", line 159, in analyze_chunks
    f = find_frequencies(data[:chunksize], freq, bits)
  File "../mingus/extra/fft.py", line 99, in find_frequencies
    p = [(abs(x) / float(n)) ** 2 * 2 for x in p[0:uniquePts]]
TypeError: slice indices must be integers or None or have an __index__ method
----------------------------------------------------------------------

+++ mingus/extra/fft.py
@@ -93,7 +93,7 @@ def find_frequencies(data, freq=44100, bits=16):
     # Fast fourier transform
     n = len(data)
     p = _fft(data)
-    uniquePts = numpy.ceil((n + 1) / 2.0)
+    uniquePts = int(numpy.ceil((n + 1) / 2.0))
 
     # Scale by the length (n) and square the value to get the amplitude
     p = [(abs(x) / float(n)) ** 2 * 2 for x in p[0:uniquePts]]
